\name{const.clust}
\alias{const.clust}
\docType{package}
\title{ Space- and time-constrained agglomerative clustering }
\description{
Compute space-constrained or time-constrained agglomerative clustering from a multivariate dissimilarity matrix.
}
\details{
\tabular{ll}{
Package: \tab const.clust\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2012-11-23\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
  Function \code{constrained.clust} carries out space-constrained
  or time-constrained agglomerative clustering from a multivariate 
  dissimilarity matrix.
}

\author{ Pierre Legendre }

\references{
Legendre, P. & L. Legendre. 1998. Numerical ecology, 2nd English edition. Elsevier Science BV, Amsterdam.
}

\keyword{ package }
